Add an option to disable bzip2 support
The explicit switch is still on by default, but this way it's possible
to explicitly disable bzip2 support even if bzip2 is available on the
system.

This is useful if bzip2 is not needed, but would automatically lead to
an unwanted dependency.

(cherry picked from commit 028c492c9d45ca29e2fbec79a203ba492016f15e)